src/backend/requirements.txt
pyhumps==3.8.0
boto3==1.26.106